Move One Tap to the client.
