+++ src/main/main.ts
@@ -38,10 +38,6 @@ import { toQueryString } from '../utils/queryString'
 import { rawRequest } from './httpClient'
 import StepUpResponse = MFA.StepUpResponse
 import MfaCredentialsResponse = MFA.CredentialsResponse
-import {randomBase64String} from "../utils/random"
-import * as OneTap from "google-one-tap"
-import {encodeToBase64} from "../utils/base64"
-import { Buffer } from 'buffer/'
 
 export { AuthResult } from './authResult'
 export { AuthOptions } from './authOptions'
@@ -67,6 +63,7 @@ export type Client = {
   loginWithCredentials: (params: LoginWithCredentialsParams) => Promise<AuthResult>
   loginWithCustomToken: (params: { token: string; auth: AuthOptions }) => Promise<void>
   loginWithPassword: (params: LoginWithPasswordParams) => Promise<AuthResult>
+  instantiateOneTap: (opts?: AuthOptions) => Promise<void>
   loginWithSocialProvider: (provider: string, options?: AuthOptions) => Promise<void | InAppBrowser>
   loginWithWebAuthn: (params: LoginWithWebAuthnParams) => Promise<AuthResult>
   logout: (params?: { redirectTo?: string; removeCredentials?: boolean }) => Promise<void>
@@ -106,7 +103,7 @@ function checkParam<T>(data: T, key: keyof T) {
   }
 }
 
-export function createClient(creationConfig: Config, oneTap = false): Client {
+export function createClient(creationConfig: Config): Client {
   checkParam(creationConfig, 'domain')
   checkParam(creationConfig, 'clientId')
 
@@ -183,8 +180,8 @@ export function createClient(creationConfig: Config, oneTap = false): Client {
     return apiClient.then(api => api.loginWithPassword(params))
   }
 
-  function idTokenRequest(provider: string, idToken: string, nonce: string, opts: AuthOptions = {}): Promise<void> {
-    return apiClient.then(api => api.idTokenRequest(provider, idToken, nonce, opts))
+  function instantiateOneTap(opts: AuthOptions = {}): Promise<void> {
+    return apiClient.then(api => api.instantiateOneTap(opts))
   }
 
   function loginWithSocialProvider(provider: string, options: AuthOptions = {}) {
@@ -309,37 +306,6 @@ export function createClient(creationConfig: Config, oneTap = false): Client {
     return apiClient.then(api => api.removeMfaEmail(params))
   }
 
-  function instanceOneTap() {
-    const GOOGLE_CLIENT_ID = "188556846346-u43frvituh5k9nore8aggnue1j1679h9.apps.googleusercontent.com"
-    const nonce = randomBase64String()
-    const binaryNonce = Buffer.from(nonce, 'utf-8')
-
-    window.crypto.subtle.digest('SHA-256', binaryNonce).then(hash => {
-      const googleIdConfiguration: OneTap.IdConfiguration = {
-        client_id: GOOGLE_CLIENT_ID,
-        callback: (response: OneTap.CredentialResponse) => idTokenRequest("google", response.credential, nonce),
-        nonce: encodeToBase64(hash),
-        // Enable auto sign-in
-        auto_select: true,
-      }
-
-      window.google.accounts.id.initialize(googleIdConfiguration)
-
-      // Activate Google One Tap
-      window.google.accounts.id.prompt()
-
-    })
-  }
-
-  if (oneTap === true) {
-    const script = document.createElement("script")
-    script.src = "https://accounts.google.com/gsi/client"
-    script.onload = instanceOneTap
-    script.async = true
-    script.defer = true
-    document.querySelector("body")?.appendChild(script)
-  }
-
   return {
     addNewWebAuthnDevice,
     checkSession,
@@ -353,6 +319,7 @@ export function createClient(creationConfig: Config, oneTap = false): Client {
     loginWithCredentials,
     loginWithCustomToken,
     loginWithPassword,
+    instantiateOneTap,
     loginWithSocialProvider,
     loginWithWebAuthn,
     logout,
